+++ pages/views/ingredients/usageData.php
@@ -23,13 +23,13 @@ $defCatClass = $settings['defCatClass'];
 
 $ing = mysqli_fetch_array(mysqli_query($conn, "SELECT id, cas,name,usage_type,noUsageLimit,byPassIFRA,flavor_use,cat1,cat2,cat3,cat4,cat5A,cat5B,cat5C,cat5D,cat6,cat7A,cat7B,cat8,cat9,cat10A,cat10B,cat11A,cat11B,cat12 FROM ingredients WHERE id = '".$_POST['ingID']."'"));
 
-$rType = searchIFRA($ing['cas'],$ing['name'],'type',$conn, $defCatClass);
-$limit = searchIFRA($ing['cas'],$ing['name'],null,$conn, 'cat'.$cats[$counter]['name']);
-if($reason = searchIFRA($ing['cas'],$ing['name'],null,$conn, $defCatClass)){
+$rType = searchIFRA($ing['cas'],$ing['name'],'type', $defCatClass);
+$limit = searchIFRA($ing['cas'],$ing['name'],null, 'cat'.$cats[$counter]['name']);
+if($reason = searchIFRA($ing['cas'],$ing['name'],null,$defCatClass)){
 	$reason = explode(' - ',$reason);
 }
 
-if($usageLimit = searchIFRA($ing['cas'],$ing['name'],null,$conn, $defCatClass)){ 
+if($usageLimit = searchIFRA($ing['cas'],$ing['name'],null,$defCatClass)){ 
 	$noLimit = 'disabled'; 
 	$byPass = 'enabled';
 }else{
@@ -37,62 +37,81 @@ if($usageLimit = searchIFRA($ing['cas'],$ing['name'],null,$conn, $defCatClass)){
 }
 ?>
 
-<h3>Usage &amp; Limits</h3>
+<h3>Usage & Limits</h3>
 <hr>
-<table width="100%" border="0">
-    <tr>
-      <td height="32">Bypass IFRA: <i rel="tip" title="Enable this to bypass IFRA values and set your own. This is not recommended though." class="pv_point_gen fas fa-info-circle"></i></td>
-      <td><input name="byPassIFRA" type="checkbox" <?php echo $byPass?:'disabled'; ?> id="byPassIFRA" value="1" <?php if($ing['byPassIFRA'] == '1'){; ?> checked="checked"  <?php } ?>/></td>
-    </tr>
-    <tr>
-        <td width="15%" height="32">No usage limit: <i rel="tip" title="This will set all values to 100% if no IFRA entries found or IFRA lookup is bypassed." class="pv_point_gen fas fa-info-circle"></i></td>
-        <td><input name="noUsageLimit" type="checkbox" <?php echo $noLimit; ?> id="noUsageLimit" value="1" <?php if($ing['noUsageLimit'] == '1'){; ?> checked="checked"  <?php } ?>/></td>
-    </tr>
-    <tr>
-        <td height="32">Flavor use:</td>
-        <td><input name="flavor_use" type="checkbox" id="flavor_use" value="1" <?php if($ing['flavor_use'] == '1'){; ?> checked="checked"  <?php } ?>/></td>
-    </tr>
-    <tr>
-        <td height="32">Usage classification:</td>
-        <td><div id="class_bypass"><?php echo $rType.' - '.$reason['1'];?></div>
+<div class="container-fluid">
+    <div class="row mb-2">
+        <div class="col-sm-2">
+            Bypass IFRA 
+            <i rel="tip" title="Enable this to bypass IFRA values and set your own. This is not recommended though." class="pv_point_gen fas fa-info-circle"></i>
+        </div>
+        <div class="col-md-4">
+            <input name="byPassIFRA" type="checkbox" <?php echo $byPass ?: 'disabled'; ?> id="byPassIFRA" value="1" <?php if($ing['byPassIFRA'] == '1'){ ?> checked="checked" <?php } ?>/>
+        </div>
+    </div>
+    <div class="row mb-2">
+        <div class="col-sm-2">
+            No usage limit
+            <i rel="tip" title="This will set all values to 100% if no IFRA entries found or IFRA lookup is bypassed." class="pv_point_gen fas fa-info-circle"></i>
+        </div>
+        <div class="col-sm-2">
+            <input name="noUsageLimit" type="checkbox" <?php echo $noLimit; ?> id="noUsageLimit" value="1" <?php if($ing['noUsageLimit'] == '1'){ ?> checked="checked" <?php } ?>/>
+        </div>
+    </div>
+    <div class="row mb-2">
+        <div class="col-sm-2">
+            Flavor use
+        </div>
+        <div class="col-sm-2">
+            <input name="flavor_use" type="checkbox" id="flavor_use" value="1" <?php if($ing['flavor_use'] == '1'){ ?> checked="checked" <?php } ?>/>
+        </div>
+    </div>
+    <div class="row mb-3">
+        <div class="col-sm-2">
+            Usage classification
+        </div>
+        <div class="col-sm-4">
+            <div id="class_bypass"><?php echo $rType.' - '.$reason['1']; ?></div>
             <div id="usage_type_bypass">
-            <select name="usage_type" id="usage_type" class="form-control">
-                <option value="1" <?php if($ing['usage_type']=="1") echo 'selected="selected"'; ?> >Recommendation</option>
-                <option value="2" <?php if($ing['usage_type']=="2") echo 'selected="selected"'; ?> >Restriction</option>
-                <option value="3" <?php if($ing['usage_type']=="3") echo 'selected="selected"'; ?> >Specification</option>
-                <option value="4" <?php if($ing['usage_type']=="4") echo 'selected="selected"'; ?> >Prohibition</option>
-            </select>
+                <select name="usage_type" id="usage_type" class="form-select">
+                    <option value="1" <?php if($ing['usage_type'] == "1") echo 'selected'; ?>>Recommendation</option>
+                    <option value="2" <?php if($ing['usage_type'] == "2") echo 'selected'; ?>>Restriction</option>
+                    <option value="3" <?php if($ing['usage_type'] == "3") echo 'selected'; ?>>Specification</option>
+                    <option value="4" <?php if($ing['usage_type'] == "4") echo 'selected'; ?>>Prohibition</option>
+                </select>
             </div>
-    </td>
-</tr>
-</table>
+        </div>
+    </div>
+</div>
 <hr />
-<table width="100%" border="0">
-<?php for($i = 0; $i < $rows/$cols; $i++) { ?>
-    <tr <?php if($rType){ ?>class="<?php echo $usageStyle[$i % 2]; ?>" <?php }?>>
-        <?php for($j=0; $j < $cols && $counter <= $rows; $j++, $counter++) {?>
-            <td align="center"><a href="#" rel="tip" title="<?php echo $cats[$counter]['description'];?>">Cat<?php echo $cats[$counter]['name'];?> %:</a></td>
-            <td>
-            <?php
-			if($ing['byPassIFRA'] == 0 &&
-            	$limit = searchIFRA($ing['cas'],$ing['name'],null,$conn, 'cat'.$cats[$counter]['name'])){
-                $limit = explode(' - ',$limit);
-			?>
-				<input name="cat<?php echo $cats[$counter]['name'];?>" type="text" class="form-control" id="cat<?php echo $cats[$counter]['name'];?>" disabled value="<?php echo number_format((float)$limit['0'],4); ?>" />
-	        <?php }else{ ?>
-                <input name="cat<?php echo $cats[$counter]['name'];?>" type="text" class="form-control" id="cat<?php echo $cats[$counter]['name'];?>" value="<?php echo number_format($ing['cat'.$cats[$counter]['name']],4); ?>" />
-            </td>
-            <?php 
-        } 
-    } 
-    ?>
-</tr>
-<?php } ?>
-</table>
+<div class="container-fluid">
+    <table class="table">
+        <?php for($i = 0; $i < $rows / $cols; $i++) { ?>
+            <tr <?php if($rType){ ?>class="<?php echo $usageStyle[$i % 2]; ?>" <?php }?>>
+                <?php for($j = 0; $j < $cols && $counter <= $rows; $j++, $counter++) { ?>
+                    <td align="center">
+                        <a href="#" rel="tip" title="<?php echo $cats[$counter]['description']; ?>">Cat<?php echo $cats[$counter]['name']; ?> %:</a>
+                    </td>
+                    <td>
+                        <?php
+                        if($ing['byPassIFRA'] == 0 && $limit = searchIFRA($ing['cas'], $ing['name'], null, 'cat'.$cats[$counter]['name'])){
+                            $limit = explode(' - ', $limit);
+                        ?>
+                            <input name="cat<?php echo $cats[$counter]['name']; ?>" type="text" class="form-control" id="cat<?php echo $cats[$counter]['name']; ?>" disabled value="<?php echo number_format((float)$limit[0], 4); ?>" />
+                        <?php } else { ?>
+                            <input name="cat<?php echo $cats[$counter]['name']; ?>" type="text" class="form-control" id="cat<?php echo $cats[$counter]['name']; ?>" value="<?php echo number_format($ing['cat'.$cats[$counter]['name']], 4); ?>" />
+                        <?php } ?>
+                    </td>
+                <?php } ?>
+            </tr>
+        <?php } ?>
+    </table>
+</div>
 <hr />
-<p> To set a category to zero, please type <strong>0.0</strong> instead of 0</p>
+<p>To set a category to zero, please type <strong>0.0</strong> instead of 0</p>
 <hr />
 <p><input type="submit" name="save" class="btn btn-primary" id="saveUsage" value="Save" /></p>
+
 <script>
 var byPassIFRA = <?=$ing['byPassIFRA']?>;
 var byPassState = '<?=$byPass?>';
@@ -115,15 +134,11 @@ $(document).ready(function() {
 			$("input[id^='cat']").prop('disabled', false);
 			$('#usage_type_bypass').show();
 			$('#class_bypass').hide();
-			//$('#tab1').hide();
-			//$('#t2').show();
 		}else if (s === false){
 			$('#noUsageLimit').prop('disabled', true);
 			$("input[id^='cat']").prop('disabled', true);
 			$('#usage_type_bypass').hide();
 			$('#class_bypass').show();
-		//	$('#tab1').show();
-		//	$('#t2').hide();
 		}
 		if (byPassState === 'disabled'){
 			$('#noUsageLimit').prop('disabled', false);